Mapping service: Don't differentiate URI and notation for from/to(Scheme)
Previously, only either the URI field or the notation field was checked in the DB query depending on whether the search string started with "http". As this is not a way to check whether a string is a URI, and because notation strings could also be mistaken as URIs, both `uri` and `notation` will be checked.